Add missing support function tests for nppi and npps
- Add tests for nppiMalloc_16s_C2, nppiMalloc_16sc (C1-C4),
  nppiMalloc_16u_C2, nppiMalloc_32f_C2, nppiMalloc_32fc (C1-C4),
  nppiMalloc_32sc (C1-C4)
- Add tests for nppsMalloc_8s, nppsMalloc_16u, nppsMalloc_16sc,
  nppsMalloc_32u, nppsMalloc_32sc, nppsMalloc_64s, nppsMalloc_64sc,
  nppsMalloc_64fc
- Fix type mismatch in test_npps_sum.cpp (size_t -> int)

+++ test/unit/npps/npps_statistics_functions/test_npps_sum.cpp
@@ -37,7 +37,7 @@ TEST_F(NPPSSumFunctionalTest, Sum_32f_BasicOperation) {
   cudaMalloc(&d_sum, sizeof(Npp32f));
 
   // 获取所需缓冲区大小
-  size_t bufferSize;
+  int bufferSize;
   NppStatus status = nppsSumGetBufferSize_32f(nLength, &bufferSize);
   ASSERT_EQ(status, NPP_NO_ERROR) << "nppsSumGetBufferSize_32f failed";
